pyproject.toml
[tool.poetry]
name = "atomic_habits"
version = "0.1.0"
description = ""
authors = ["Konstantin <kos26193@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.0"
pillow = "^12.1.0"
djangorestframework = "^3.16.1"
python-dotenv = "^1.2.1"
django-filter = "^25.2"
djangorestframework-simplejwt = "^5.5.1"
drf-yasg = "^1.21.14"
stripe = "^14.3.0"
forex-python = "^1.9.2"
celery = "^5.6.2"
redis = "^7.1.0"
eventlet = "^0.40.4"
django-celery-beat = "^2.8.1"
poetry-core = "^2.3.1"
django-cors-headers = "^4.3.0"
psycopg2-binary = "^2.9.11"
psycopg = "^3.3.2"

[tool.poetry.group.lint.dependencies]
black = "^26.1.0"
flake8 = "^7.3.0"
isort = "^7.0.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

